REFACTOR-#3861: move 'SignalActor' class into task_wrapper.py
